Add tools section
