+++ libs/design-system/src/lib/Components/Flips/index.tsx
@@ -27,10 +27,7 @@ export default function Flips({ flips }: FlipsProps) {
       </p>
 
       <div className="mb-6">
-        <TabMenu
-          tabs={['Open FLIPs', 'Good places to start']}
-          onTabChange={() => null}
-        />
+        <TabMenu tabs={['Tools', ' Concepts']} onTabChange={() => null} />
       </div>
 
       <FlipCellHeader />
