+++ libs/design-system/src/lib/Components/index.tsx
@@ -15,3 +15,4 @@ export { default as SocialLinksSignup } from './SocialLinksSignup';
 export { default as SporksCard } from './SporksCard';
 export { default as TabMenu } from './TabMenu';
 export { default as Tag } from './Tag';
+export { default as ToolsAndConcepts } from './ToolsAndConcepts';
