+++ libs/design-system/src/lib/Pages/HomePage/index.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import {
   LandingHeaderHome,
   SocialLinksSignup,
@@ -6,86 +5,19 @@ import {
   LinkCard2Column,
   LinkCard3Column,
   Flips,
+  ToolsAndConcepts,
 } from '../../Components';
-import { ReactComponent as UseCaseIcon } from '../../../../images/content/use-cases.svg';
-import { ReactComponent as EcosystemIcon } from '../../../../images/content/ecosystem.svg';
-import { ReactComponent as SDKIcon } from '../../../../images/content/sdk.svg';
 import { FlipCellProps } from '../../Components/FlipCell';
+import { ToolCardProps } from '../../Components/ToolCard';
+import { LinkCard3ColumnProps } from '../../Components/LinkCard3Column';
 
 export type HomePageProps = {
   flips: FlipCellProps[];
+  tools: ToolCardProps[];
+  threeColumnItems: LinkCard3ColumnProps;
 };
 
-const HomePage = ({ flips }) => {
-  const threeColumnItems = [
-    {
-      title: 'Quickstart',
-      description:
-        'A package used to interact with user wallets and the Flow blockchain.',
-      icon: <UseCaseIcon />,
-      links: [
-        {
-          title: 'Quickstart tutorial',
-          href: '#tutorial1',
-          tags: ['tutorial'],
-        },
-        {
-          title: 'Name of a tutorial',
-          href: '#tutorial2',
-          tags: ['tutorial'],
-        },
-        {
-          title: 'Name of another tutorial',
-          href: '#tutorial3',
-          tags: ['tutorial'],
-        },
-      ],
-    },
-    {
-      title: 'Guides & Tutorials',
-      description:
-        'An up to 3-line blurb here describing the section lorem ipsum dolor sit amet proin.',
-      icon: <EcosystemIcon />,
-      links: [
-        {
-          title: 'Guide 1',
-          href: '#tutorial1',
-          tags: ['tutorial'],
-        },
-        {
-          title: 'Guide 2',
-          href: '#tutorial2',
-        },
-        {
-          title: 'An external link',
-          href: 'https://www.onflow.org',
-          tags: ['tutorial', 'external'],
-        },
-      ],
-    },
-    {
-      title: 'Smart Contracts',
-      description: 'Smart contracts description.',
-      icon: <SDKIcon />,
-      links: [
-        {
-          title: 'Name of a Smart Contract tutorial',
-          href: '#tutorial1',
-          tags: ['tutorial'],
-        },
-        {
-          title: 'Name of a tutorial',
-          href: '#tutorial2',
-          tags: ['tag1', 'tag2', 'tag3', 'tag4'],
-        },
-        {
-          title: "View all SDK's",
-          href: '#sdks',
-        },
-      ],
-    },
-  ];
-
+const HomePage = ({ flips, tools, threeColumnItems }) => {
   return (
     <div>
       <LandingHeaderHome
@@ -103,6 +35,10 @@ const HomePage = ({ flips }) => {
       />
 
       <LinkCard3Column items={threeColumnItems as any} />
+
+      <div className="p-6 my-20">
+        <ToolsAndConcepts tools={tools} />
+      </div>
 
       <Flips flips={flips} />
